backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
chromadb
python-dotenv
langchain-text-splitters
pypdf
python-docx